requirements.txt
absl-py==0.12.0
apache-beam==2.31.0rc1
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
avro-python3==1.9.1
backcall==0.2.0
bidict==0.21.2
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
dill==0.3.1.1
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.0
fastavro==1.4.1
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.52.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.3.1
importlib-resources==5.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
netifaces==0.10.6
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.0rc1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.0.0
promise==2.3
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.17.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
python-engineio==3.8.2.post1
python-slugify==5.0.2
python-socketio==4.2.1
pytz==2021.1
pywin32==301
pyyaml==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
-e git+https://github.com/Sailesh-2209/autonav.git@305b5038b4fd7d73e6e9b82dbe8632a94e794c5d#egg=slim&subdirectory=object detection\dependencies\apimodel\research\slim
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.1.0
tensorflow-model-optimization==0.6.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.5.0
tf-slim==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
zipp==3.4.1
